# docs/swagger.yaml
openapi: 3.0.3
info:
  title: MotoMarket API
  description: |
    API REST pour la marketplace MotoMarket - Vente et achat de motos, pièces détachées et accessoires.
    
    ## Authentification
    La plupart des routes nécessitent un token JWT dans le header `Authorization: Bearer <token>`.
    
    ## Pagination
    Les listes retournent un objet paginé avec `data`, `page`, `limit`, `total`, `totalPages`.
    
    ## Upload Photos
    Utiliser `multipart/form-data` pour les uploads. Formats acceptés: JPG, JPEG, PNG, WebP.
    Taille max configurable côté serveur.
    
    ## Types d'annonces
    - **MOTORCYCLE** : Motos complètes
    - **PART** : Pièces détachées (freins, moteurs, pneus...)
    - **ACCESSORY** : Accessoires (casques, blousons, gants...)
  version: 1.0.0
  contact:
    name: MotoMarket
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Développement local
  - url: https://api.motomarket.mg/api
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ListingType:
      type: string
      enum: [MOTORCYCLE, PART, ACCESSORY]
      description: "Type d'annonce : MOTORCYCLE (moto), PART (pièce détachée), ACCESSORY (accessoire)"
    ListingStatus:
      type: string
      enum: [DRAFT, PENDING_REVIEW, ACTIVE, SOLD, PAUSED, EXPIRED, REJECTED, DELETED]
    MotorcycleCategory:
      type: string
      enum: [CROSS, ROUTE, ROADSTER, SCOOTER, TRAIL, CUSTOM, AUTRE]
      description: "Catégorie de moto (uniquement pour les annonces de type MOTORCYCLE)"
    FuelType:
      type: string
      enum: [ESSENCE, DIELECTRIQUE, HYBRIDE, ELECTRIQUE]
    TransmissionType:
      type: string
      enum: [MANUELLE, AUTOMATIQUE, SEMI_AUTO]
    ConditionType:
      type: string
      enum: [NEUF, TRES_BON, BON, USAGE, A_REFORMER]
    SellerType:
      type: string
      enum: [INDIVIDUAL, PROFESSIONAL]
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 250
        totalPages:
          type: integer
          example: 13

    # ── Auth ──
...
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logo:
          type: string
          nullable: true
        modelsCount:
          type: integer
    MotorcycleModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        name:
          type: string
        brand:
          $ref: '#/components/schemas/Brand'
...
    PartCategory:
      type: object
      description: "Catégorie de pièce détachée"
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Freins"
        slug:
          type: string
          example: "freins"
        icon:
          type: string
          nullable: true
    AccessoryCategory:
      type: object
      description: "Catégorie d'accessoire"
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Casque"
        slug:
          type: string
          example: "casque"
        icon:
          type: string
          nullable: true

    # ── Annonce (Listing) ──
    Listing:
      type: object
      description: "Annonce unifiée (moto, pièce ou accessoire)"
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ListingType'
        sellerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: integer
          description: "Prix en Ariary (MGA)"
        isPriceNegotiable:
          type: boolean
        condition:
          $ref: '#/components/schemas/ConditionType'
        city:
          type: string
        district:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ListingStatus'
        isFeatured:
          type: boolean
        publishedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        viewsCount:
          type: integer
        images:
          type: array
          items:
            $ref: '#/components/schemas/ListingImage'
        seller:
          $ref: '#/components/schemas/SellerSummary'
        # Champs spécifiques MOTO
        brand:
          $ref: '#/components/schemas/Brand'
          nullable: true
        model:
          $ref: '#/components/schemas/MotorcycleModel'
          nullable: true
        motorcycleCategory:
          $ref: '#/components/schemas/MotorcycleCategory'
          nullable: true
        year:
          type: integer
          nullable: true
        mileage:
          type: integer
          nullable: true
        engineCc:
          type: integer
          nullable: true
        fuel:
          $ref: '#/components/schemas/FuelType'
          nullable: true
        transmission:
          $ref: '#/components/schemas/TransmissionType'
          nullable: true
        maintenanceInfo:
          type: string
          nullable: true
        papersInfo:
          type: string
          nullable: true
        modifications:
          type: string
          nullable: true
        # Champs spécifiques PIÈCE
        partCategory:
          $ref: '#/components/schemas/PartCategory'
          nullable: true
        compatibleBrands:
          type: string
          nullable: true
        partReference:
          type: string
          nullable: true
        # Champs spécifiques ACCESSOIRE
        accessoryCategory:
          $ref: '#/components/schemas/AccessoryCategory'
          nullable: true
        accessorySize:
          type: string
          nullable: true
        accessoryColor:
          type: string
          nullable: true
        accessoryBrand:
          type: string
          nullable: true
        favoritesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListingSummary:
      type: object
      description: "Version allégée pour les listes"
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ListingType'
        title:
          type: string
        price:
          type: integer
        condition:
          $ref: '#/components/schemas/ConditionType'
        city:
          type: string
        status:
          $ref: '#/components/schemas/ListingStatus'
        isFeatured:
          type: boolean
        primaryImage:
          $ref: '#/components/schemas/ListingImage'
        seller:
          $ref: '#/components/schemas/SellerSummary'
        # Motos
        brand:
          $ref: '#/components/schemas/Brand'
          nullable: true
        model:
          $ref: '#/components/schemas/MotorcycleModel'
          nullable: true
        year:
          type: integer
          nullable: true
        mileage:
          type: integer
          nullable: true
        engineCc:
          type: integer
          nullable: true
        # Pièces
        partCategory:
          $ref: '#/components/schemas/PartCategory'
          nullable: true
        # Accessoires
        accessoryCategory:
          $ref: '#/components/schemas/AccessoryCategory'
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListingCreateRequest:
      type: object
      required: [type, title, description, price, condition, city]
      properties:
        type:
          $ref: '#/components/schemas/ListingType'
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
          minLength: 20
          maxLength: 5000
        price:
          type: integer
          minimum: 0
        isPriceNegotiable:
          type: boolean
          default: false
        condition:
          $ref: '#/components/schemas/ConditionType'
        city:
          type: string
        district:
          type: string
          nullable: true
        # Motos
        brandId:
          type: string
          format: uuid
        modelId:
          type: string
          format: uuid
        motorcycleCategory:
          $ref: '#/components/schemas/MotorcycleCategory'
        year:
          type: integer
          minimum: 1950
          maximum: 2030
        mileage:
          type: integer
          minimum: 0
        engineCc:
          type: integer
          minimum: 50
          maximum: 2000
        fuel:
          $ref: '#/components/schemas/FuelType'
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        maintenanceInfo:
          type: string
          nullable: true
        papersInfo:
          type: string
          nullable: true
        modifications:
          type: string
          nullable: true
        # Pièces
        partCategoryId:
          type: string
          format: uuid
        compatibleBrands:
          type: string
        partReference:
          type: string
        # Accessoires
        accessoryCategoryId:
          type: string
          format: uuid
        accessorySize:
          type: string
        accessoryColor:
          type: string
        accessoryBrand:
          type: string
    ListingUpdateRequest:
      type: object
      description: "Tous les champs sont optionnels"
      properties:
        type:
          $ref: '#/components/schemas/ListingType'
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
          minLength: 20
          maxLength: 5000
        price:
          type: integer
          minimum: 0
        isPriceNegotiable:
          type: boolean
        condition:
          $ref: '#/components/schemas/ConditionType'
        city:
          type: string
        district:
          type: string
        brandId:
          type: string
          format: uuid
        modelId:
          type: string
          format: uuid
        motorcycleCategory:
          $ref: '#/components/schemas/MotorcycleCategory'
        year:
          type: integer
        mileage:
          type: integer
        engineCc:
          type: integer
        fuel:
          $ref: '#/components/schemas/FuelType'
        transmission:
          $ref: '#/components/schemas/TransmissionType'
        partCategoryId:
          type: string
          format: uuid
        compatibleBrands:
          type: string
        partReference:
          type: string
        accessoryCategoryId:
          type: string
          format: uuid
        accessorySize:
          type: string
        accessoryColor:
          type: string
        accessoryBrand:
          type: string
        favoritesCount:
          type: integer
        viewsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SellerSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        avatar:
          type: string
          nullable: true
        isVerifiedSeller:
          type: boolean
        sellerType:
          $ref: '#/components/schemas/SellerType'
        businessName:
          type: string
          nullable: true
        averageRating:
          type: number
          format: float
    StatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ListingStatus'

    # ── Images ──
    ListingImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        position:
          type: integer
        isPrimary:
          type: boolean
    ImageUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        position:
          type: integer
        isPrimary:
          type: boolean

    # ── Favorites ──
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [relevance, newest, price_asc, price_desc, mileage_asc, year_desc]
        default: relevance
paths:
...
  /listings:
    get:
      tags: [Listings]
      summary: Liste des annonces (recherche + filtres)
      operationId: listListings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ListingType'
          description: "Filtrer par type : MOTORCYCLE, PART ou ACCESSORY"
        - name: search
          in: query
          schema:
            type: string
          description: "Recherche texte libre"
        - name: brandId
          in: query
          schema:
            type: string
            format: uuid
        - name: modelId
          in: query
          schema:
            type: string
            format: uuid
        - name: partCategoryId
          in: query
          schema:
            type: string
            format: uuid
          description: "Filtrer par catégorie de pièce"
        - name: accessoryCategoryId
          in: query
          schema:
            type: string
            format: uuid
          description: "Filtrer par catégorie d'accessoire"
        - name: minPrice
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: minYear
          in: query
          schema:
            type: integer
        - name: maxYear
          in: query
          schema:
            type: integer
        - name: maxMileage
          in: query
          schema:
            type: integer
        - name: engineCc
          in: query
          schema:
            type: integer
        - name: motorcycleCategory
          in: query
          schema:
            $ref: '#/components/schemas/MotorcycleCategory'
        - name: fuel
          in: query
          schema:
            $ref: '#/components/schemas/FuelType'
        - name: transmission
          in: query
          schema:
            $ref: '#/components/schemas/TransmissionType'
        - name: condition
          in: query
          schema:
            $ref: '#/components/schemas/ConditionType'
        - name: city
          in: query
          schema:
            type: string
        - name: sellerType
          in: query
          schema:
            type: string
            enum: [INDIVIDUAL, PROFESSIONAL]
        - name: verifiedSeller
          in: query
          schema:
            type: boolean
        - name: isFeatured
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListingSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Listings]
      summary: Créer une annonce
      operationId: createListing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        '201':
          description: Annonce créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          description: Données invalides
        '401':
          description: Non authentifié
  /listings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Listings]
      summary: Détail d'une annonce
      operationId: getListing
      responses:
        '200':
          description: Détail complet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '404':
          description: Annonce non trouvée

    put:
      tags: [Listings]
      summary: Modifier une annonce
      operationId: updateListing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdateRequest'
      responses:
        '200':
          description: Annonce modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '403':
          description: Pas le propriétaire
        '404':
          description: Annonce non trouvée

    delete:
      tags: [Listings]
      summary: Supprimer une annonce
      operationId: deleteListing
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Supprimée
        '403':
          description: Pas le propriétaire
  /listings/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Listings]
      summary: Changer le statut d'une annonce
      operationId: updateListingStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: Statut mis à jour
        '403':
          description: Non autorisé
  /listings/{id}/images:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Listings]
      summary: Ajouter des photos
      operationId: uploadImages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  minItems: 1
                  maxItems: 15
      responses:
        '201':
          description: Photos uploadées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageUploadResponse'
        '403':
          description: Pas le propriétaire
  /listings/{id}/images/{imageId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: imageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Listings]
      summary: Supprimer une photo
      operationId: deleteImage
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Photo supprimée
        '403':
          description: Pas le propriétaire

  # ═══════════════════════════════════════════════
  # FAVORITES
  # ═══════════════════════════════════════════════
...
  /admin/listings:
    get:
      tags: [Admin]
      summary: Toutes les annonces
      operationId: adminListListings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ListingStatus'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Annonces
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListingSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '403':
          description: Admin uniquement
  /admin/listings/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Admin]
      summary: Modérer une annonce (admin)
      operationId: adminUpdateStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: Statut mis à jour
        '403':
          description: Admin uniquement
...